Fix compiler warnings surfaced by clang-cl -Weverything on Windows CI
- Control.c, DebugViewController.c, TableView.c: explicit casts to the
  target enum type at three int-to-enum assignment/call sites.
- Font.c: (void) init to silence -Wunused-variable when NDEBUG strips
  the assert that was init's only use; TTF_Init() itself still always runs.
- Renderer.c: explicitly brace the position/texcoord sub-objects (and
  zero the unused color field) in MVC_Vertex literals -- pushDrawArrays
  always overwrites per-vertex color from its own parameter.
- Slider.c: explicit double cast before the float/double subtraction.
- Text.c, Window.c: assign a float-returning call to an intermediate
  local before casting to int, instead of casting the call result
  directly.

Verified: all touched files compile clean; the only remaining failure
is the pre-existing, unrelated `ld: framework 'Objectively' not found`
in this environment.

+++ Sources/ObjectivelyMVC/Slider.c
@@ -188,7 +188,7 @@ static bool captureEvent(Control *self, const SDL_Event *event) {
     if (self->state & ControlStateHighlighted) {
       if (frame.w) {
 
-        const double fraction = (event->motion.x - frame.x) / (double) frame.w;
+        const double fraction = (double) (event->motion.x - frame.x) / (double) frame.w;
         double value = this->min + (this->max - this->min) * clamp(fraction, 0.0, 1.0);
 
         if (this->snapToStep && this->step) {
